move utility functions to plotutil
